internal/controller/blobstorage_controller.go
package controller

import (
	"context"

	"k8s.io/apimachinery/pkg/runtime"
	ctrl "sigs.k8s.io/controller-runtime"
	"sigs.k8s.io/controller-runtime/pkg/client"
	"sigs.k8s.io/controller-runtime/pkg/log"

	k8sv1beta1 "github.com/migueleliasweb/s3-bucket-k8s-operator/api/v1beta1"
)

// BlobstorageReconciler reconciles a Blobstorage object
type BlobstorageReconciler struct {
	client.Client
	Scheme *runtime.Scheme
}

// +kubebuilder:rbac:groups=k8s.migueleliasweb.io,resources=blobstorages,verbs=get;list;watch;create;update;patch;delete
// +kubebuilder:rbac:groups=k8s.migueleliasweb.io,resources=blobstorages/status,verbs=get;update;patch
// +kubebuilder:rbac:groups=k8s.migueleliasweb.io,resources=blobstorages/finalizers,verbs=update

// Reconcile is part of the main kubernetes reconciliation loop which aims to
// move the current state of the cluster closer to the desired state.
// TODO(user): Modify the Reconcile function to compare the state specified by
// the Blobstorage object against the actual cluster state, and then
// perform operations to make the cluster state reflect the state specified by
// the user.
//
// For more details, check Reconcile and its Result here:
// - https://pkg.go.dev/sigs.k8s.io/controller-runtime@v0.19.0/pkg/reconcile
func (r *BlobstorageReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
	_ = log.FromContext(ctx)

	// TODO(user): your logic here

	return ctrl.Result{}, nil
}

// SetupWithManager sets up the controller with the Manager.
func (r *BlobstorageReconciler) SetupWithManager(mgr ctrl.Manager) error {
	return ctrl.NewControllerManagedBy(mgr).
		For(&k8sv1beta1.Blobstorage{}).
		Complete(r)
}
